+++ MK4duo/Configuration_Temperature.h
@@ -95,6 +95,8 @@
 
 // Enable this for support DHT sensor for temperature e Humidity DHT11, DHT21 or DHT22.
 //#define DHT_SENSOR
+// Set Type DHT 11 for DHT11, 21 for DHT21, 22 for DHT22
+#define DHT_TYPE 11
 
 //These 2 defines help to calibrate the AD595 sensor in case you get wrong temperature measurements.
 //The measured temperature is defined as "actualTemp = (measuredTemp * TEMP_SENSOR_AD595_GAIN) + TEMP_SENSOR_AD595_OFFSET"